Adds exclude from recents option
Adds a setting to exclude the app from the recent apps list.

This enhances user privacy by preventing the app from appearing in the recent apps overview, especially when sensitive information might be displayed.

The feature is implemented by using ActivityManager.appTasks.setExcludeFromRecents(boolean).

This setting only hides the app when returning to the home screen via the back key.